+++ packages/compiler-core/__tests__/scopeId.spec.ts
@@ -70,18 +70,15 @@ describe('scopeId compiler support', () => {
     expect(ast.helpers).toContain(PUSH_SCOPE_ID)
     expect(ast.helpers).toContain(POP_SCOPE_ID)
     expect(ast.hoists.length).toBe(2)
-    expect(code).toMatch(
-      [
-        `_pushScopeId("test")`,
-        `const _hoisted_1 = /*#__PURE__*/_createElementVNode("div", null, "hello", ${genFlagText(
-          PatchFlags.HOISTED
-        )})`,
-        `const _hoisted_2 = /*#__PURE__*/_createElementVNode("div", null, "world", ${genFlagText(
-          PatchFlags.HOISTED
-        )})`,
-        `_popScopeId()`
-      ].join('\n')
-    )
+    ;[
+      `const _withScopeId = n => (_pushScopeId("test"),n=n(),_popScopeId(),n)`,
+      `const _hoisted_1 = /*#__PURE__*/ _withScopeId(() => /*#__PURE__*/_createElementVNode("div", null, "hello", ${genFlagText(
+        PatchFlags.HOISTED
+      )}))`,
+      `const _hoisted_2 = /*#__PURE__*/ _withScopeId(() => /*#__PURE__*/_createElementVNode("div", null, "world", ${genFlagText(
+        PatchFlags.HOISTED
+      )}))`
+    ].forEach(c => expect(code).toMatch(c))
     expect(code).toMatchSnapshot()
   })
 })